Remove gridding and replace with pyvisgrid methods, Add pyvisgrid as github dependency (prelim), Fix typos, Move dataset simulations to dataset module
