+++ src/pyvisgen/dataset/__init__.py
@@ -1,0 +1,4 @@
+from . import utils
+from .data_set import SimulateDataSet
+
+__all__ = ["SimulateDataSet", "utils"]
