+++ src/pyvisgen/simulation/__init__.py
@@ -1,5 +1,4 @@
 from .array import Array
-from .data_set import SimulateDataSet
 from .observation import Baselines, Observation, ValidBaselineSubset
 from .scan import angular_distance, calc_beam, calc_fourier, integrate, jinc, rime
 from .visibility import Polarization, Visibilities, generate_noise, vis_loop
@@ -9,7 +8,6 @@ __all__ = [
     "Baselines",
     "Observation",
     "Polarization",
-    "SimulateDataSet",
     "ValidBaselineSubset",
     "Visibilities",
     "angular_distance",
