+++ tests/test_utils.py
@@ -23,7 +23,7 @@ def test_read_config():
         "bandwidths",
         "corrupted",
         "noisy",
-        "sensitivty_cut",
+        "sensitivity_cut",
         "polarization",
         "pol_delta",
         "pol_amp_ratio",
